JavaScript Image Slider
Now we need to handle the Next and Prev button clicks and call the corresponding functions to change the image.
HTML:

Try It Yourself

The num variable holds the current image. The next and previous button clicks are handled by their corresponding functions, which change the source of the image to the next/previous image in the array.
We have created a functioning image slider!